+++ examples/setup-temp/tailwind.config.ts
@@ -3,5 +3,6 @@ import { type Config } from "tailwindcss";
 export default {
   content: [
     "{routes,islands,components}/**/*.{ts,tsx,js,jsx}",
+    "../../src/styles.ts",
   ],
 } satisfies Config;